fix: change type of update process input
The ProcessInputSchema was being used in update process,
which is supposed to limit users', this is wrong, since it
doesn't allow the server to make changes.
To limit the changes from users, I think it would be best to use the
ProcessInputSchema in the server action exposed to the client.