Ms2/signin providers (#320)
* typo

* feat(ms2/signin): discord sign in

* feat(ms2/signin): twitter login

twitter doesn't provide an email if you don't link your privacy policy
and terms of service. As we don't have them, we can't get the email from
twitter right now, so I had to make the email optional in the user
schema.

* refactor(ms2/signin): use providers in production

* fix(ms2): updated env variables

* typo

* fix(ms2/sysadmins): don't make guests admins

* feat(ms2/sysadmins): consider fist sysadmin when guest signs in

* fix(ms2/sysadmins): use global rules for sysadmin abilities

* Fixed: Missing await in fixture for the base page (#322)

* feat(ms2/signin): profile for twitter provider

* feat(ms2/signin): profile for discord provider

* fix(ms2): optional email

this is needed because Twitter doesn't return an email. This commit
should definitely be reverted and the type made not optionional once it
does.

* ms2: todo message

---------

Co-authored-by: jjoderis <58050428+jjoderis@users.noreply.github.com>
Co-authored-by: Kai Rohwer <rohwer.kai@icloud.com>