refactor(ms2): get processes takes in an spaceId and an ability
This change was needed, because the function was redundant and was
implemented in a way that caused it to fail.
It is redundant to use a userId and an ability, because both tell you
for which user we're supposed to get the processes.
And it fails because, it returns the processes where the creator was the
user, this is wrong, because it doesn't take into consideration
abilities.

For me is better that the function takes in a spaceId, and returns all
processes for a spaceId, and if you want to get the processes for a user
you can provide his ability.