ms2/miscellanous fixes (#644)
* fix(ms2/share-modal): allow copy to clipboard with selection

* fix(ms2/space-select): more consistent behaviour

fixed two issues:

- Previously when the select was clicked and it happened to fall in the
  anchor tag, the user would be redirected to the process page of the
  current space. That is not what is expected when opening the space
  select
- When the click selected the option but was just outside the anchor tag,
  nothing happened, although the select showed that the space was
  changed

* feat(ms2/share-modal): preview of images copied to clipboard

* feat(ms2/executions-dashboard): skeleton for loading animation

* fix(ms2/blockly-editor): remove duplicate effect + update ref when changing to read only

There was an issue where the editor is mounted and is editable, if you
switch to a version, the new editor didn't update the blocklyEditorRef.
After the switch when you try to update the editor, the blocklyEditorRef
contains a ref to the old unmounted editor, this causes an error

* fix(ms2/auth): dev user sign no longer creates admin user

This is because the admin user is now created through the seed file.

* refactor(ms2/auth): less verbose error logging

Only log the entire error for relevant errors. Previously on any sort of
error the log was very verbose, for example when a user typed the wrong
password we got a very long log.

* feat(ms2/modeler): warning when closing tab when there are unsaved changes

* fix(ms2/modeler): expose immediate and asyncImmediate

* feat(ms2/process-list): use wysiwyg editor/view for processes