feat(engine/script-execution): routes handling in request-server
Now you can define a route as follows in the script task /<route>
and to call it the route is
/running-processes/<process-instance-id>/<route>
This allows the script task to specify routes without knowing the
process instance id, we could provide it, but to me it seems like
an extra step.
The downside is that the route defined in the scripttask is not exactly
the one that has to be called.