fix(ms2/import-process): check filename without path
There was a change to the way images are stored: we now store the entire
filePath (corresponding to the bucket/file-system) in the src of the
image, instead of computing it every time we fetch an image.
The previous code wasn't expecting the src of the image to be the path.