fix importing and deploying processes (#662)
* fix(ms2/import-process): check filename without path

There was a change to the way images are stored: we now store the entire
filePath (corresponding to the bucket/file-system) in the src of the
image, instead of computing it every time we fetch an image.
The previous code wasn't expecting the src of the image to be the path.

* fix(engine/resource-route): return mime type text

Since no mime type was specified, the returned value was sent to the MS
as JSON which caused an error when parsing it.