fix(ms2/user-task): get images used in production build
The `getUsedImagesFromJson` function used the DisplayName of components,
which is changed it the production build, thus when saving a User Task
no images where detected and the references were not updated.
Later on when trying to deploy a process this led to an error:
images are looked up by by their filename and also with a reference
(a version reference), which doesn't exist.